+++ authpass/lib/cloud_storage/cloud_storage_helper.dart
@@ -1,15 +1,17 @@
 import 'dart:convert';
 
+import 'package:authpass/cloud_storage/cloud_storage_provider.dart';
 import 'package:authpass/env/_base.dart';
 import 'package:authpass/utils/path_util.dart';
 import 'package:biometric_storage/biometric_storage.dart';
 
 /// Common functionality shared across all cloud storages,
 /// right now simply storing of oauth tokens.
-class CloudStorageHelper {
+class CloudStorageHelper implements CloudStorageHelperBase {
   CloudStorageHelper(this.env, this.pathUtil);
 
   final Env env;
+  @override
   final PathUtil pathUtil;
   BiometricStorageFile _storageFile;
 
@@ -33,10 +35,12 @@ class CloudStorageHelper {
     await file.write(json.encode(dataMap));
   }
 
+  @override
   Future<String> loadCredentials(String cloudStorageId) async {
     return (await _readDataMap())[cloudStorageId] as String;
   }
 
+  @override
   Future<void> saveCredentials(String cloudStorageId, String data) async {
     final dataMap = await _readDataMap();
     dataMap[cloudStorageId] = data;