+++ authpass/lib/env/development.dart
@@ -1,8 +1,9 @@
 import 'package:authpass/env/_base.dart';
+import 'package:authpass/env/env.dart';
 
 Future<void> main() async => await Development().start();
 
-class Development extends Env {
+class Development extends EnvAppBase {
   Development() : super(EnvType.development);
 
   @override