+++ authpass/lib/env/fdroid.dart
@@ -1,8 +1,9 @@
 import 'package:authpass/env/_base.dart';
+import 'package:authpass/env/env.dart';
 
 Future<void> main() async => await FDroid().start();
 
-class FDroid extends Env {
+class FDroid extends EnvAppBase {
   FDroid() : super(EnvType.production);
 
   @override