+++ authpass/lib/env/production.dart
@@ -1,10 +1,11 @@
 import 'package:authpass/env/_base.dart';
+import 'package:authpass/env/env.dart';
 
 import 'secrets.dart';
 
 Future<void> main() async => await Production().start();
 
-class Production extends Env {
+class Production extends EnvAppBase {
   Production() : super(EnvType.production);
 
   @override