+++ authpass/lib/env/test_driver.dart
@@ -1,8 +1,9 @@
 import 'package:authpass/env/_base.dart';
+import 'package:authpass/env/env.dart';
 
 Future<void> main() async => await TestDriverEnv().start();
 
-class TestDriverEnv extends Env {
+class TestDriverEnv extends EnvAppBase {
   TestDriverEnv() : super(EnvType.production);
 
   @override